+++ modules/sqidCore/src/processing/ops/source.cpp
@@ -35,6 +35,8 @@ namespace sqid
 		_lastUpdateTime( 0 ),
 		_sampleRate( 0.0f ),
 		_sampleCntr( 0 ),
+		_frameRate( 0.0f ),
+		_frameCntr( 0 ),
 		_dataRate( 0 ),
 		_dataCntr( 0 ),
 		_statsTimeAccu( 0.0f ),
@@ -95,7 +97,8 @@ namespace sqid
 		if( !sf )
 			return false;
 
-		_sampleCntr++;
+		_sampleCntr += sf->size();
+		_frameCntr++;
 		_dataCntr += sf->size() * sizeof( float );
 
 		_lastUpdateTime = getAppTime();
@@ -120,9 +123,11 @@ namespace sqid
 		if( _statsTimeAccu > 1.0f )
 		{
 			_sampleRate = _sampleCntr / _statsTimeAccu;
+			_frameRate = _frameCntr / _statsTimeAccu;
 			_dataRate = _dataCntr / _statsTimeAccu;
 
 			_sampleCntr = 0;
+			_frameCntr = 0;
 			_dataCntr = 0;
 
 			_statsTimeAccu = 0.0f;
@@ -189,6 +194,7 @@ namespace sqid
 		}
 
 		ImGui::Text( "%.02f sps", _sampleRate );
+		ImGui::Text( "%.02f fps", _frameRate );
 		ImGui::Text( "%.02f kbps", ( _dataRate << 3 ) / 1024.0f );
 		if( _sourceDesc.size() )
 			ImGui::Text( _sourceDesc.c_str() );